% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Advertising data for examples}
\description{
Advertising expenditure in three different media for
some product with corresponding sales of this product.
Fabricated data for the purpose of examples in the analytics course.
}
