% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrAnalytics-package.R
\docType{package}
\name{jrAnalytics-pacakge}
\alias{jrAnalytics-pacakge}
\alias{jrAnalytics}
\alias{jrAnalytics-pacakge-package}
\title{Introduction to analytics in R by Jumping Rivers}
\description{
Functions and data sets used in R courses
}
\keyword{package}
