% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zodiac}
\alias{zodiac}
\title{Zodiac signs of CEOs}
\description{
Numbers of successful business personal by zodiac sign.
We are interested in whether there is any association between star sign and success.
}
